Refactor motes tooltip
