+++ src/main/java/de/hysky/skyblocker/skyblock/item/tooltip/ItemTooltip.java
@@ -129,13 +129,6 @@ public class ItemTooltip {
         }
 
 
-        if (TooltipInfoType.MOTES.isTooltipEnabledAndHasOrNullWarning(internalID)) {
-            lines.add(Text.literal(String.format("%-20s", "Motes Price:"))
-                    .formatted(Formatting.LIGHT_PURPLE)
-                    .append(getMotesMessage(TooltipInfoType.MOTES.getData().get(internalID).getAsInt(), count)));
-        }
-
-
         if (TooltipInfoType.MUSEUM.isTooltipEnabledAndHasOrNullWarning(internalID) && !bazaarOpened) {
             String itemCategory = TooltipInfoType.MUSEUM.getData().get(internalID).getAsString();
             String format = switch (itemCategory) {
@@ -327,26 +320,6 @@ public class ItemTooltip {
         String priceStringTotal = String.format(Locale.ENGLISH, "%1$,.1f", price * count);
         MutableText message = Text.literal(priceStringTotal + " Coins ").formatted(Formatting.DARK_AQUA);
         message.append(Text.literal("(" + priceString + " each)").formatted(Formatting.GRAY));
-
-        return message;
-    }
-
-    private static Text getMotesMessage(int price, int count) {
-        float motesMultiplier = SkyblockerConfigManager.get().otherLocations.rift.mcGrubberStacks * 0.05f + 1;
-
-        // Calculate the total price
-        int totalPrice = price * count;
-        String totalPriceString = String.format(Locale.ENGLISH, "%1$,.1f", totalPrice * motesMultiplier);
-
-        // If count is 1, return a simple message
-        if (count == 1) {
-            return Text.literal(totalPriceString.replace(".0", "") + " Motes").formatted(Formatting.DARK_AQUA);
-        }
-
-        // If count is greater than 1, include the "each" information
-        String eachPriceString = String.format(Locale.ENGLISH, "%1$,.1f", price * motesMultiplier);
-        MutableText message = Text.literal(totalPriceString.replace(".0", "") + " Motes ").formatted(Formatting.DARK_AQUA);
-        message.append(Text.literal("(" + eachPriceString.replace(".0", "") + " each)").formatted(Formatting.GRAY));
 
         return message;
     }
